Fixed a problem where the scroll position got hosed each time the user clicked the play button.
In order to make `max-width:100%;` work without squashing images, you can't set the `height`.  This means that the container div just takes the height of its contents.  When an element is removed and then added, this means that the div briefly has height 0.

If the user is looking at an image at the bottom of a tall page, when the container div shrinks, the user is force-scrolled higher on the page, and the updated image (paused or playing) is shown below the user's view.

Swapping `appendChild/removeChild` for `replaceChild` doesn't fix the problem.  So, I used one div to react to changes in width, a second div to maintain the proper aspect ratio, and a third div to hold the actual content.  I got the idea from [here](http://stackoverflow.com/a/10441480/1153071).  In addition to fixing the scrolling bug, it also looks nicer - there's no flicker artifact when switching from playing to stopped, even when the images have been shrunk to fit page width.